triggers/requirements.txt
requests==2.18.4
python_dateutil==2.6.1
Flask==0.12.2
redis==2.10.6
